[2888] Ensure the retrieval of properly loaded view models
Bug: https://github.com/eclipse-sirius/sirius-web/issues/2888
