+++ CHANGELOG.adoc
@@ -92,6 +92,7 @@ Additional changes to this interface will propably occur in the near future in o
 - https://github.com/eclipse-sirius/sirius-web/issues/2826[#2826] [diagram] Fix invalid reference position used for nodes with a depth superior to 2.
 - https://github.com/eclipse-sirius/sirius-web/issues/2857[#2857] [form] Fix default reference widget bugs.
 Now, for a multi-valued feature, the values are properly displayed and the delete is properly handled.
+- https://github.com/eclipse-sirius/sirius-web/issues/2888[#2888] [view] Ensure that the retrieval of the view models returns properly loaded models
 
 === New Features
 - https://github.com/eclipse-sirius/sirius-web/issues/2413[#2413] Add a debug mode for React Flow diagrams
