+++ packages/sirius-web/backend/sirius-web-services/src/main/java/org/eclipse/sirius/web/services/editingcontext/DynamicRepresentationDescriptionService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2023 Obeo.
+ * Copyright (c) 2021, 2024 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -12,36 +12,18 @@
  *******************************************************************************/
 package org.eclipse.sirius.web.services.editingcontext;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.UUID;
-import java.util.stream.Stream;
 
-import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EPackage;
-import org.eclipse.emf.ecore.resource.Resource;
-import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl;
-import org.eclipse.emf.ecore.util.ECrossReferenceAdapter;
 import org.eclipse.emf.edit.domain.EditingDomain;
-import org.eclipse.sirius.components.emf.ResourceMetadataAdapter;
-import org.eclipse.sirius.components.emf.services.EditingContext;
-import org.eclipse.sirius.components.emf.services.JSONResourceFactory;
 import org.eclipse.sirius.components.representations.IRepresentationDescription;
 import org.eclipse.sirius.components.view.View;
-import org.eclipse.sirius.components.view.ViewPackage;
 import org.eclipse.sirius.components.view.emf.IViewConverter;
-import org.eclipse.sirius.emfjson.resource.JsonResource;
-import org.eclipse.sirius.web.persistence.entities.DocumentEntity;
-import org.eclipse.sirius.web.persistence.repositories.IDocumentRepository;
 import org.eclipse.sirius.web.services.editingcontext.api.IDynamicRepresentationDescriptionService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.eclipse.sirius.web.services.editingcontext.api.IViewLoader;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
 
 /**
@@ -52,19 +34,14 @@ import org.springframework.stereotype.Service;
 @Service
 public class DynamicRepresentationDescriptionService implements IDynamicRepresentationDescriptionService {
 
-    private final Logger logger = LoggerFactory.getLogger(DynamicRepresentationDescriptionService.class);
-
-    private final IDocumentRepository documentRepository;
-
-    private final EPackage.Registry ePackageRegistry;
+    private final IViewLoader viewLoader;
 
     private final IViewConverter viewConverter;
 
     private final boolean isStudioDefinitionEnabled;
 
-    public DynamicRepresentationDescriptionService(IDocumentRepository documentRepository, EPackage.Registry ePackageRegistry, IViewConverter viewConverter, @Value("${org.eclipse.sirius.web.features.studioDefinition:false}") boolean isStudioDefinitionEnabled) {
-        this.documentRepository = Objects.requireNonNull(documentRepository);
-        this.ePackageRegistry = Objects.requireNonNull(ePackageRegistry);
+    public DynamicRepresentationDescriptionService(IViewLoader viewLoader, IViewConverter viewConverter, @Value("${org.eclipse.sirius.web.features.studioDefinition:false}") boolean isStudioDefinitionEnabled) {
+        this.viewLoader = Objects.requireNonNull(viewLoader);
         this.viewConverter = Objects.requireNonNull(viewConverter);
         this.isStudioDefinitionEnabled = isStudioDefinitionEnabled;
     }
@@ -73,34 +50,14 @@ public class DynamicRepresentationDescriptionService implements IDynamicRepresen
     public List<IRepresentationDescription> findDynamicRepresentationDescriptions(String editingContextId, EditingDomain editingDomain) {
         List<IRepresentationDescription> dynamicRepresentationDescriptions = new ArrayList<>();
         if (this.isStudioDefinitionEnabled) {
-            List<View> views = new ArrayList<>();
-            List<EPackage> accessibleEPackages = this.getAccessibleEPackages(editingDomain);
-            ResourceSet resourceSet = this.createResourceSet(this.ePackageRegistry);
-            this.loadStudioColorPalettes(resourceSet);
+            List<View> views = this.viewLoader.load();
 
-            this.documentRepository.findAllByType(ViewPackage.eNAME, ViewPackage.eNS_URI).forEach(documentEntity -> {
-                Resource resource = this.loadDocumentAsEMF(documentEntity, resourceSet);
-                views.addAll(this.getViewDefinitions(resource).toList());
-            });
+            List<EPackage> accessibleEPackages = this.getAccessibleEPackages(editingDomain);
             this.viewConverter.convert(views, accessibleEPackages).stream()
                     .filter(Objects::nonNull)
                     .forEach(dynamicRepresentationDescriptions::add);
         }
         return dynamicRepresentationDescriptions;
-    }
-
-    private void loadStudioColorPalettes(ResourceSet resourceSet) {
-        ClassPathResource classPathResource = new ClassPathResource("studioColorPalettes.json");
-        URI uri = URI.createURI(EditingContext.RESOURCE_SCHEME + ":///" + UUID.nameUUIDFromBytes(classPathResource.getPath().getBytes()));
-        Resource resource = new JSONResourceFactory().createResource(uri);
-        try (var inputStream = new ByteArrayInputStream(classPathResource.getContentAsByteArray())) {
-            resourceSet.getResources().add(resource);
-            resource.load(inputStream, null);
-            resource.eAdapters().add(new ResourceMetadataAdapter("studioColorPalettes"));
-        } catch (IOException exception) {
-            this.logger.warn("An error occured while loading document studioColorPalettes.json: {}.", exception.getMessage());
-            resourceSet.getResources().remove(resource);
-        }
     }
 
     private List<EPackage> getAccessibleEPackages(EditingDomain editingDomain) {
@@ -110,27 +67,5 @@ public class DynamicRepresentationDescriptionService implements IDynamicRepresen
                 .filter(EPackage.class::isInstance)
                 .map(EPackage.class::cast)
                 .toList();
-    }
-
-    private Stream<View> getViewDefinitions(Resource resource) {
-        return resource.getContents().stream().filter(View.class::isInstance).map(View.class::cast);
-    }
-
-    private Resource loadDocumentAsEMF(DocumentEntity documentEntity, ResourceSet resourceSet) {
-        JsonResource resource = new JSONResourceFactory().createResourceFromPath(documentEntity.getId().toString());
-        resourceSet.getResources().add(resource);
-        try (var inputStream = new ByteArrayInputStream(documentEntity.getContent().getBytes())) {
-            resource.load(inputStream, null);
-        } catch (IOException | IllegalArgumentException exception) {
-            this.logger.warn(exception.getMessage(), exception);
-        }
-        return resource;
-    }
-
-    private ResourceSet createResourceSet(EPackage.Registry packageRegistry) {
-        ResourceSet resourceSet = new ResourceSetImpl();
-        resourceSet.eAdapters().add(new ECrossReferenceAdapter());
-        resourceSet.setPackageRegistry(packageRegistry);
-        return resourceSet;
     }
 }
