+++ packages/sirius-web/backend/sirius-web-services/src/main/java/org/eclipse/sirius/web/services/editingcontext/api/IViewLoader.java
@@ -1,0 +1,26 @@
+/*******************************************************************************
+ * Copyright (c) 2024 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.sirius.web.services.editingcontext.api;
+
+import java.util.List;
+
+import org.eclipse.sirius.components.view.View;
+
+/**
+ * Used to load properly view models.
+ *
+ * @author sbegaudeau
+ */
+public interface IViewLoader {
+    List<View> load();
+}
