+++ packages/sirius-web/backend/sirius-web-services/src/main/java/org/eclipse/sirius/web/services/representations/ViewRepresentationDescriptionSearchServiceParameters.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 Obeo.
+ * Copyright (c) 2023, 2024 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -21,7 +21,7 @@ import org.eclipse.sirius.components.view.emf.diagram.IDiagramIdProvider;
 import org.eclipse.sirius.components.view.emf.form.IFormIdProvider;
 import org.eclipse.sirius.components.view.emf.task.IDeckIdProvider;
 import org.eclipse.sirius.components.view.emf.task.IGanttIdProvider;
-import org.eclipse.sirius.web.persistence.repositories.IDocumentRepository;
+import org.eclipse.sirius.web.services.editingcontext.api.IViewLoader;
 import org.springframework.stereotype.Service;
 
 /**
@@ -32,9 +32,7 @@ import org.springframework.stereotype.Service;
 @Service
 public final class ViewRepresentationDescriptionSearchServiceParameters {
 
-    private final IDocumentRepository documentRepository;
-
-    private final EPackage.Registry ePackageRegistry;
+    private final IViewLoader viewLoader;
 
     private final IDiagramIdProvider diagramIdProvider;
 
@@ -48,12 +46,11 @@ public final class ViewRepresentationDescriptionSearchServiceParameters {
 
     private final IObjectService objectService;
 
-    public ViewRepresentationDescriptionSearchServiceParameters(IDocumentRepository documentRepository, EPackage.Registry ePackageRegistry, IURLParser urlParser, IObjectService objectService,
+    public ViewRepresentationDescriptionSearchServiceParameters(IViewLoader viewLoader, EPackage.Registry ePackageRegistry, IURLParser urlParser, IObjectService objectService,
             ViewRepresentationIdParameters representationIdParameters) {
         Objects.requireNonNull(representationIdParameters);
         this.urlParser = Objects.requireNonNull(urlParser);
-        this.documentRepository = Objects.requireNonNull(documentRepository);
-        this.ePackageRegistry = Objects.requireNonNull(ePackageRegistry);
+        this.viewLoader = Objects.requireNonNull(viewLoader);
         this.diagramIdProvider = Objects.requireNonNull(representationIdParameters.getDiagramIdProvider());
         this.formIdProvider = Objects.requireNonNull(representationIdParameters.getFormIdProvider());
         this.ganttIdProvider = Objects.requireNonNull(representationIdParameters.getGanttIdProvider());
@@ -61,12 +58,8 @@ public final class ViewRepresentationDescriptionSearchServiceParameters {
         this.objectService = Objects.requireNonNull(objectService);
     }
 
-    public IDocumentRepository getDocumentRepository() {
-        return this.documentRepository;
-    }
-
-    public EPackage.Registry getEPackageRegistry() {
-        return this.ePackageRegistry;
+    public IViewLoader getViewLoader() {
+        return this.viewLoader;
     }
 
     public IDiagramIdProvider getDiagramIdProvider() {
